refactor: split useLoadDetails into useDetailsData and useDetailsOperations hooks
